add controllers, services and exceptions

+++ src/main/java/com/platzi/market/domain/service/ProductService.java
@@ -1,9 +1,9 @@
 package com.platzi.market.domain.service;
 
 import com.platzi.market.domain.dto.ProductDto;
+import com.platzi.market.domain.exceptions.CustomEmptyResultException;
 import com.platzi.market.domain.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,7 +12,7 @@ import java.util.Optional;
 @Service
 public class ProductService {
 
-    private ProductRepository productRepository;
+    private final ProductRepository productRepository;
 
     @Autowired
     public ProductService(ProductRepository productRepository) {
@@ -35,11 +35,30 @@ public class ProductService {
         return productRepository.save(product);
     }
 
+    // TODO PROBAR Y REVISAR. SOLUCION PROPUESTA POR UN ALUMNO Y CHATGPT. REVISAR VALIDACIONES
+    // https://chatgpt.com/share/a073ce0a-2443-493c-969d-04763e964a5f
+    public Optional<ProductDto> update(long id, ProductDto productDto) {
+        Optional<ProductDto> optProductDto = productRepository.getProduct(id);
+        if (optProductDto.isEmpty()) {
+            return Optional.empty();
+        } else {
+            ProductDto productToUpdate = optProductDto.get();
+            if (productDto.getName() != null) productToUpdate.setName(productDto.getName());
+            if (productDto.getCategoryId() != null) productToUpdate.setCategoryId(productDto.getCategoryId());
+            if (productDto.getPrice() != null) productToUpdate.setPrice(productDto.getPrice());
+            if (productDto.getStock() != null) productToUpdate.setStock(productDto.getStock());
+            if (productDto.getActive() != null) productToUpdate.setActive(productDto.getActive());
+            return Optional.of(productRepository.save(productToUpdate));
+        }
+    }
+
     public boolean delete(long id) {
         try {
+            if (!productRepository.existsById(id)) throw new CustomEmptyResultException(id);
             productRepository.delete(id);
             return true;
-        } catch (EmptyResultDataAccessException e) {
+        } catch (CustomEmptyResultException e) {
+            // TODO REGISTER ERROR IN LOGS
             return false;
         }
     }
